+++ risan.h
@@ -1,4 +1,4 @@
 ﻿#pragma once
 #include <stdio.h>
 #include <stdbool.h>
-int risan();
+long long risan();
